test update

+++ test/test.py
@@ -46,17 +46,28 @@ class TestExperiment(unittest.TestCase):
         r = os.system("diff test-en-nl/test-en-nl.phrasetable.targetdump test-en-nl.phrasetable.targetdump.ok")
         self.assertEqual(r,0)
 
-    def test004_trainfiles(self):
-        """Verifying training instances"""
+    def test004_trainfiles_X(self):
+        """Verifying training instances for experts"""
         self.assertEqual(  len(list(glob.glob("test-en-nl/classifierdata-XI2l1r1/*.train"))), 2)
         r = os.system("diff test-en-nl/classifierdata-XI2l1r1/bank.train bank.train.ok")
         self.assertEqual(r,0)
         r = os.system("diff test-en-nl/classifierdata-XI2l1r1/the+bank.train the+bank.train.ok")
         self.assertEqual(r,0)
 
+    def test004_trainfiles_M(self):
+        """Verifying training instances for monolithic system"""
+        self.assertEqual(  len(list(glob.glob("test-en-nl/classifierdata-MI2l1r1/*.train"))), 1)
+        r = os.system("diff test-en-nl/classifierdata-MI2l1r1/train.train train.train.ok")
+        self.assertEqual(r,0)
+
 if __name__ == '__main__':
+        if not os.path.exists("test.py"):
+            print("Please run the test from the test/ directory",file=sys.stderr)
+            sys.exit(2)
+
         print("Cleanup",file=sys.stderr)
-        os.system("rm -Rf test-en-nl > /dev/null 2> /dev/null")
+        if os.path.exists("test-en-nl"):
+            os.system("rm -Rf test-en-nl")
         os.system("mkdir test-en-nl")
         os.system("cp test-en-nl.phrasetable test-en-nl/")
